renamed package to bigquery_views_manager
