+++ bigquery_views_manager/delete_views_or_tables.py
